requirements.txt
torch
torchvision
pillow
tqdm
scikit-learn